Update to gazelle 0.39.1
This fixes a few of the TODOs in the WORKSPACE file and changes a lot
about the popular_repos tests.

This commit also provides a default value of
`--java_runtime_version=remotejdk_21` just for the bazel_test in
lcov_coverage_test.go so that contributors who don't
have a JAVA_HOME on the host system can still get the
lcov_coverage_test.go to pass by default.

Oh, and it also fixes a TODO in the repository's `.bazelrc` file.