Wrap generate_imported_dylib.sh in genrules
As a contributor to the project, I expect to be able to clone the
repository, go to the master branch and run `bazel test //...` and
have it work.

But, before this change, just like the CI system, contributors were
also expected to run `./tests/core/cgo/generate_imported_dylib.sh`
manually. Well ... no more!