Support the go1.21 `testing.Testing()` function
This change adds in the necessary flags in `go test` to allow
[`testing.Testing`] to work as expected.  According to the [upstream
source], it is set with a `-X` option in the `go` tooling, so this
change makes `rules_go` operate accordingly.

This should be safe for older Go versions too, given that the link
command [docs](https://pkg.go.dev/cmd/link) say that the `-X` arg is
not "effective"" when the variable is not defined in the code.

Tests are also provided.

Fixes #3686.

[`testing.Testing`]: https://pkg.go.dev/testing#Testing
[upstream source]: https://cs.opensource.google/go/go/+/refs/tags/go1.21.0:src/testing/testing.go;l=647-661