Keep order for PATH components
The starlark `dict` already will return items in the order in which
they where inserted on iteration. There is no need to explicitly
sort the items. But it does break use cases, where the user has
explicitly defined a order of `PATH` components. As for example
is possible with the `env` attribute of `cc_args()` with the new
`rule_cc` based toolchains.