chore(releaser): allow upgrade-dep to take a specific version
We need to update repositories.bzl to stay in sync with go.mod, but
ideally we don't want to avoid updates in go.mod. This requires
updating the tool to be able to request a specific version rather than
just updating to the latest.

When requesting a package at a specific version, the comment should not
claim that it was "latest" as of today.