Revert "Fail when expected files are not produced by protoc (#4287)" (#4324)
**What type of PR is this?**
Bug fix


**What does this PR do? Why is it needed?**
It's common for a `go_proto_library` to include proto files with and
without service definitions at the same time. In this case, the gRPC
plugins are needed, but some gRPC plugins don't generate the grpc.pb.go
files if there is no service definition.

Partially reverting #4287 to restore the previous behavior of creating
an empty file

**Which issues(s) does this PR fix?**

Fixes #4317

**Other notes for review**

---------

Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>