Remove shadow analyzer from TOOLS_NOGO
The shadow analyzer issues many false positives which are not in keeping
with Go best practices. Even the upstream documentation for this
analyzer admits this and claims that it is experimental. It should not
be included in the default list of analyzers provided by nogo (users can
still opt-in to it if they want it).