Provide `DefaultInfo` on Go toolchain rules
This makes it easier to deal with these toolchains in language-agnostic aspects and helpers (in particular, bazel_env.bzl), that propagate to toolchain targets.

Even though genrules now support `toolchain_type`s in their `toolchains`, this does *not* allow Go SDKs to be used in genrules as they don't expose the path to the Go binary via a Make variable. This is intentional as we don't want to allow running `go` in `genrule`s.