coverageredesign: Always use atomic mode
coverageredesign enables the use of "runtime/coverage"
which are APIs designed to enable expose runtime
coverage data in  non test programs.

"runtime/coverage" explicitly only works with "atomic" covermode
ref: https://github.com/golang/go/blob/6fbad4be75e7746512bbe55794694ed788ea5c5b/src/internal/coverage/cfile/apis.go#L39

In this light, always invoke `go tool cover` with covermode atomic
as it would be unxpected for test and non-test coverage to behave
differently.

ref #3513