Add host_compatible flag to extension
This controls whether the tag is considered for the host compatible
toolchain. This is necessary if any of the tags point to a toolchain
that should not be used to build host tools.

For example, we use a compiler that _requires_ cgo. For the code in the
repository this works because cgo is enabled and we use a hermetic
clang. But some tools like `builder` and `gazelle` either disable cgo or
do not pass in the compiler.