coverageredesign: Map line directive file names to exec relative paths
coverageredesign (https://github.com/bazel-contrib/rules_go/pull/4397)
honors line directives while the world before coverageredesign did not.

This is relevant in situations where you want coverage for generated
code to map to source code in a user's workspace.

An example of this is a go_library rule which takes a dependency
on a code generator which generates a replacement file for a
user source file. The replacement file is run in the final binary
but the user wants to see covarege for the source file that was
an input to the generated replacement file.

```
go_library(
    name = "go_default_library",
    srcs = [
        "batch.go",
        "consumers.go",
        "edges.go",
        "token.go",
        ":codegen",
    ],
)

codegen(name = "codegen, ...)
```

The Go compiler, because of Bazel, only knows about the generated code
but coverage is only relevant for the source file that the
code generator replaces.

Given a generated file
`bazel-out/k8-fastbuild/bin/src/example.org/gen_/report_gen.go`
nocoverageredesign would emit that file path in coverage files.

With coverageredesign, if a line directive in the generated code
changes the file name to hello.go (e.g. //line hello.go:10) then
the path in the runtime emitted coverage file would be
`example.org/gen_/hello.go` which is not execution root relative
and would be ignored by Bazel.

The prevent this code from breaking in coverage redesign, we need
to map the line directive file name to the execution root relative
filepath that Bazel respects to get coverage for these types
of files.

This change extends the "srcPathMapping" trick used in
https://github.com/bazel-contrib/rules_go/pull/4397/files
to honor a mapping of line directive file names to map back to
exec root relative paths that Bazel understands and requires
for lcov mode.