remove .vscode and gitignore it
The .vscode directory isn't meant to be checked in like this, and causes
churn that vscode users have to remove before sending PRs. We remove it,
add it to .gitignore so future folks don't accidentally check it in.
Folks can check out the more general Editor setup wiki page for
instructions.

Fixes #4497.